requirements.txt
discord.py>=2.3.0
docker>=7.0.0